+++ src/main/java/com/healthshop/domain/order/Order.java
@@ -1,4 +1,4 @@
-package com.healthshop.domain;
+package com.healthshop.domain.order;
 
 import jakarta.persistence.*;
 import lombok.Getter;
