+++ src/main/java/com/healthshop/domain/order/OrderStatus.java
@@ -1,4 +1,4 @@
-package com.healthshop.domain;
+package com.healthshop.domain.order;
 
 public enum OrderStatus {
     COMPLETE, CANCEL
