+++ src/main/java/com/healthshop/domain/order/PaymentMethod.java
@@ -1,4 +1,4 @@
-package com.healthshop.domain;
+package com.healthshop.domain.order;
 
 public enum PaymentMethod {
     CREDIT_CARD, ACCOUNT_TRANSFER
