+++ src/main/java/com/healthshop/healthshop/MemberRepository.java
@@ -1,6 +1,6 @@
 package com.healthshop.healthshop;
 
-import com.healthshop.domain.Member;
+import com.healthshop.domain.member.Member;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import org.springframework.stereotype.Repository;
